% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{baseline_BSI}
\alias{baseline_BSI}
\title{Process Baseline BSI data}
\usage{
baseline_BSI(baseline)
}
\arguments{
\item{baseline}{Data frame containing the baseline survey data}
}
\value{
A data frame containing the BSI and PTSD scores
}
\description{
This function will calculate the depression, anxiety,
somatic, and global subscales of the BSI, and the
PTSD score of the PCL-5 from the baseline data set.
}
