% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{baseline_SES}
\alias{baseline_SES}
\title{Process Baseline SES data}
\usage{
baseline_SES(baseline)
}
\arguments{
\item{baseline}{Data frame containing the baseline survey data}
}
\value{
A data frame containing binned income and education values
}
\description{
This function will recode and return income and education
brackets from the baseline survey.
}
\details{
Income is coded as 1 = < 10k, 2 = 10-30k, 3 = 30-50k, 4 = 50-80k,
5 = 80-120k, 6 = 120-160k, 7 = 160-200k, 8 = 200-250k, 9 = 250k+.

Education is coded as 1 = less than high school/GED, 2 = high school/GED, 3 = trade
school or partial college, 4 = 2-year degree, 5 = 4-year degree, 6 = graduate degree
}
