% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_all_timepoints}
\alias{get_all_timepoints}
\title{Provides timepoint completion info for every timepoint}
\usage{
get_all_timepoints(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
dataframe with each ID, and a column for each timepoint with date of completion
}
\description{
Provides timepoint completion info for every timepoint
}
