% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_baseline}
\alias{get_baseline}
\title{Process Baseline data}
\usage{
get_baseline(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the
full data for all subjects who completed
the baseline COPE survey.
}
