% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_beck}
\alias{get_beck}
\title{Process Beck Depression Scores}
\usage{
get_beck(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
Beck Depression Inventory
}
