% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_cbq}
\alias{get_cbq}
\title{Process 42m Childhood Behavior Questionnaire Data}
\usage{
get_cbq(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the
Childhood Behavior Questionnaire dataset for 42 months
}
