% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_chaos}
\alias{get_chaos}
\title{Process CHAOS data}
\usage{
get_chaos(token, timepoint = "infant_9months_arm_1", all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Redcap event name}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
CHAOS scale
}
