% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_child_dob}
\alias{get_child_dob}
\title{Generates Child DOB Info}
\usage{
get_child_dob(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
data frame of record id and most recent/accurate child_dob variable & IDs to check manually
}
\description{
Parses through all child DOB variable, removes errors, returns final date & age in days
For IDs with no DOB data, also returns due date & expected age based on that
For IDs with only 2 DOB entries, both different, the earliest will be chosen & a message printed indicating that ID should be checked
}
