% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_data}
\alias{get_data}
\title{Load a Matrix}
\usage{
get_data(token = token, form = form, raw_v_label = "raw", form_complete = T)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{form}{Name of the survey to be downloaded}

\item{raw_v_label}{Whether raw data or labels are requested}

\item{form_complete}{Whether complete or all surveys are downloaded (default is T)}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download any requested survey from REDCap.
The following missing data codes in REDCap are replaced
with NA: -888,888, 8888, -999, 999, 9999. Only complete
surveys are returned.
}
