% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_ecbq}
\alias{get_ecbq}
\title{Process Early Childhood Behavior Questionnaire Data}
\usage{
get_ecbq(token, all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the
Early Childhood Behavior Questionnaire dataset for 30 months
}
