% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_eefq}
\alias{get_eefq}
\title{Process Early Executives Functions Questionnaire}
\usage{
get_eefq(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
EEFQ scale
}
