% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_epds}
\alias{get_epds}
\title{Process EPDS postpartum depression data}
\usage{
get_epds(token, timepoint = "infant_6months_arm_1", all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the total
summed scores for the EPDS. Only total scores, and
not cutoff values, are returned.
}
