% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_expected_invites}
\alias{get_expected_invites}
\title{Generates Expected Invite Data}
\usage{
get_expected_invites(token, timepoint = timepoint, max_date = "none")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{The COPE timepoint in numeric format (6, 9, 12, 18, 30, 42)}

\item{max_date}{The max date you wish to pull data for (e.g. '2024-06-01'), default is all future invites}
}
\value{
A list with 1) data frame of all ids, invite date, month, longitudinal status, 2) stacked bar chart of counts by month, 3) data frame of counts by month
}
\description{
Lists all expected invite dates/months for specified timepoints and time frames
}
