% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_field}
\alias{get_field}
\title{Pull a Specifield Survey Field}
\usage{
get_field(token = token, field = field, raw_v_label = "raw")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{field}{Name of the specific field to be downloaded}

\item{raw_v_label}{Whether raw data or labels are requested}
}
\value{
A data frame for the completed record_ids, redcap_event_name and field
}
\description{
This function will return redcap_event_name and field data for each event
}
