% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_ibq}
\alias{get_ibq}
\title{Process IBQ data}
\usage{
get_ibq(token, timepoint = "infant_6months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and process
the effortful control, negative affect,
and surgency subscales of the infant
behavior record.
}
