% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_infant_care}
\alias{get_infant_care}
\title{Process Infant Care data}
\usage{
get_infant_care(token, timepoint)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return parenting attention
and infant reward subscales from the ICQ. NOTE:
these measures were defined by DMW and are not validated subscales
in the original survey.
}
