% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_ious27}
\alias{get_ious27}
\title{Process IOUS Scores}
\usage{
get_ious27(token, timepoint = "infant_30months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap event name (default is 'infant_30months_arm_1')}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
IOUS scale
}
