% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_mom_dob}
\alias{get_mom_dob}
\title{Generates Mom DOB Info}
\usage{
get_mom_dob(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
data frame of record id and most recent/accurate mom_dob variable & IDs to check manually
}
\description{
Parses through all mom DOB variable, removes errors, returns final date & current age
For IDs with only 2 DOB entries, both different, the earliest will be chosen & a message printed indicating that ID should be checked
}
