% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_mpas}
\alias{get_mpas}
\title{Process Maternal Postnatal Attachment Scale}
\usage{
get_mpas(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return average
scores for the pleasure and quality subscales
of the MPAS.
}
