% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_parent_reward}
\alias{get_parent_reward}
\title{Process Parenting Reward Data}
\usage{
get_parent_reward(token, timepoint = "infant_3months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the average
parenting reward scores.
}
