% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_parent_strain}
\alias{get_parent_strain}
\title{Process Parenting Role Strain Data}
\usage{
get_parent_strain(token, timepoint = "infant_6months_arm_1", all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return average
parenting role strain scores, only for the questions
probing CURRENT strain and NOT past strain
}
