% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_phq}
\alias{get_phq}
\title{Process Patient Health Questionnaire Scores}
\usage{
get_phq(token, timepoint = "infant_9months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap event name (default is 'infant_9months_arm_1')}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
PHQ scale
}
