% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_prenatal_covid_records}
\alias{get_prenatal_covid_records}
\title{Find prenatal COVID+ moms}
\usage{
get_prenatal_covid_records(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame with record IDs of moms self-reporting as having covid during pregnancy
}
\description{
This function will find infant record IDs of moms who self-reported having covid during pregnancy
}
