% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_resilience}
\alias{get_resilience}
\title{Process Resilience Data}
\usage{
get_resilience(token, timepoint = "infant_9months_arm_1", all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap event name for timepoint you want to pull}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
resilience survey
}
