% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_srq}
\alias{get_srq}
\title{Process Social Reward Questionnaire}
\usage{
get_srq(token, all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
SRQ
}
