% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_stai}
\alias{get_stai}
\title{Process State-Trait Anxiety Inventory Scores}
\usage{
get_stai(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
STAI
}
