% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Generates REDCap Survey Data}
\usage{
get_survey_data(token, form = form, event = event)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{form}{Name of the survey to be downloaded}

\item{event}{Event name e.g. infant_30months_arm_1}
}
\value{
Data frame with record_ids and survey information for particular form
}
\description{
Returns a data frame with contact info, completion status, access codes etc. for every ID for a given survey
}
