% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_visit_data}
\alias{get_visit_data}
\title{Generates data for specified timepoint}
\usage{
get_visit_data(token, timepoint = "infant_6months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap_event_name for the timepoint you wish to pull}
}
\value{
dataframe with visit data for that timepoint
}
\description{
This function returns a dataframe with infant dob, date, and age for a specified event
}
