% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_visit_n}
\alias{get_visit_n}
\title{Generating N for Specified Timepoint}
\usage{
get_visit_n(token, timepoint = "infant_6months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{redcap_event_name for the event you wish to pull}
}
\value{
Numeric variable for that event
}
\description{
This function with return a numeric variable for the number of people who have completed a particular event
}
