% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_asq}
\alias{get_asq}
\title{Process ASQ data}
\usage{
get_asq(token, all_items = F)
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{all_items}{Boolean - whether to pull raw item scores as well as the subscales}
}
\value{
A data frame for the processed ASQ data
}
\description{
This function will download and compute scores for
all ASQ subscales at 6 months. Only total scores, and
not cut off values, are returned.
}
