% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_chaos}
\alias{get_chaos}
\title{Process CHAOS data}
\usage{
get_chaos(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and compute total scores for the
CHAOS scale
}
