% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_infant_care}
\alias{get_infant_care}
\title{Process Infant Care data}
\usage{
get_infant_care(token, timepoint = "infant_6months_arm_1")
}
\arguments{
\item{token}{Unique REDCap token ID}

\item{timepoint}{Survey timepoint branch requested}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the
Infant Care Questionnaire dataset
}
