% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopeData.R
\name{get_thq}
\alias{get_thq}
\title{Process Trauma History data}
\usage{
get_thq(token)
}
\arguments{
\item{token}{Unique REDCap token ID}
}
\value{
A data frame for the completed surveys
}
\description{
This function will download and return the total number
of traumatic events reported by subjects.
}
